+++ api/routes/agency.ts
@@ -7,7 +7,7 @@ import * as Default from '../lib/limits.js';
 
 export const AgencyResponse = Type.Object({
     id: Type.Integer(),
-    name: Type.String()
+    name: Type.String(),
 });
 
 export default async function router(schema: Schema, config: Config) {