+++ cloudformation/lib/pmtiles.js
@@ -105,10 +105,10 @@ export default {
             }
         },
         PMTilesApiMap: {
-           Type: 'AWS::ApiGateway::BasePathMapping',
-           Properties: {
-               DomainName: cf.ref('PMTilesApiDomain'),
-               RestApiId: cf.ref('PMTilesLambdaAPI'),
+            Type: 'AWS::ApiGateway::BasePathMapping',
+            Properties: {
+                DomainName: cf.ref('PMTilesApiDomain'),
+                RestApiId: cf.ref('PMTilesLambdaAPI')
             }
         },
         PMTilesLambdaAPI: {
@@ -202,7 +202,7 @@ export default {
         },
         PMTilesAPICNAME: {
             Description: 'PMTiles API CNAME target',
-            Value: cf.join([ cf.getAtt('PMTilesApiDomain', 'RegionalDomainName'), '.']),
+            Value: cf.join([cf.getAtt('PMTilesApiDomain', 'RegionalDomainName'), '.']),
             Export: {
                 Name: cf.join([cf.stackName, '-pmtiles-api-cname-target'])
             }